# controller/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: go-admin
  description: go-admin
paths:
  /api/v1/users:
    get:
      parameters:
        - name: username
          in: query
          schema:
            type: string
        - name: name
          in: query
          style: form
          schema:
            type: string
        - name: status
          in: query
          style: form
          schema:
            type: string
        - name: current
          in: query
          schema:
            type: integer
          required: true
        - name: pageSize
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: delete
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: delete
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  /api/v1/roles:
    get:
      parameters:
        - name: name
          in: query
          style: form
          schema:
            type: string
        - name: status
          in: query
          style: form
          schema:
            type: string
        - name: current
          in: query
          schema:
            type: integer
          required: true
        - name: pageSize
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/roles/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: delete
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: delete
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    
    User:
      properties:
        username:
          type: string
          maxLength: 64
        password:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 64
        email:
          type: string
          maxLength: 128
        phone:
          type: string
          maxLength: 32
        remark:
          type: string
          maxLength: 1024
        status:
          type: string
          enum:
            - activated
            - frozen
        role:
          items:
            $ref: '#/components/schemas/UserRole'
          type: array
        created:
          type: string
        updated:
          type: string
      required:
        - username
        - password
      type: object
      
    UserRole:
      properties:
        id:
          type: integer
          format: int32
        user_id:
          type: integer
          format: int32
        role_id:
          type: integer
          format: int32
        role_name:
          type: string
        created:
          type: string
        updated:
          type: string
      type: object
      
    Role:
      properties:
        name:
          type: string
          maxLength: 128
        code:
          type: string
          maxLength: 32
        description:
          type: string
        sequence:
          type: integer
        status:
          type: string
          enum:
            - enabled
            - disabled
        created:
          type: string
        updated:
          type: string
        menu:
          items:
            $ref: '#/components/schemas/RoleMenu'
          type: array
      required:
        - name
        - code
      type: object
      
    RoleMenu:
      properties:
        id:
          type: integer
          format: int32
        role_id:
          type: integer
          format: int32
        menu_id:
          type: integer
          format: int32
        created:
          type: string
        updated:
          type: string
      type: object