# controller/openapi_types.gen.go
// Package controller provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.1-0.20240331212514-80f0b978ef16 DO NOT EDIT.
package controller

// Defines values for RoleStatus.
const (
	Disabled RoleStatus = "disabled"
	Enabled  RoleStatus = "enabled"
)

// Defines values for UserStatus.
const (
	Activated UserStatus = "activated"
	Frozen    UserStatus = "frozen"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// Role defines model for Role.
type Role struct {
	Code        string      `json:"code"`
	Created     *string     `json:"created,omitempty"`
	Description *string     `json:"description,omitempty"`
	Menu        *[]RoleMenu `json:"menu,omitempty"`
	Name        string      `json:"name"`
	Sequence    *int        `json:"sequence,omitempty"`
	Status      *RoleStatus `json:"status,omitempty"`
	Updated     *string     `json:"updated,omitempty"`
}

// RoleStatus defines model for Role.Status.
type RoleStatus string

// RoleMenu defines model for RoleMenu.
type RoleMenu struct {
	Created *string `json:"created,omitempty"`
	Id      *int32  `json:"id,omitempty"`
	MenuId  *int32  `json:"menu_id,omitempty"`
	RoleId  *int32  `json:"role_id,omitempty"`
	Updated *string `json:"updated,omitempty"`
}

// User defines model for User.
type User struct {
	Created  *string     `json:"created,omitempty"`
	Email    *string     `json:"email,omitempty"`
	Name     *string     `json:"name,omitempty"`
	Password string      `json:"password"`
	Phone    *string     `json:"phone,omitempty"`
	Remark   *string     `json:"remark,omitempty"`
	Role     *[]UserRole `json:"role,omitempty"`
	Status   *UserStatus `json:"status,omitempty"`
	Updated  *string     `json:"updated,omitempty"`
	Username string      `json:"username"`
}

// UserStatus defines model for User.Status.
type UserStatus string

// UserRole defines model for UserRole.
type UserRole struct {
	Created  *string `json:"created,omitempty"`
	Id       *int32  `json:"id,omitempty"`
	RoleId   *int32  `json:"role_id,omitempty"`
	RoleName *string `json:"role_name,omitempty"`
	Updated  *string `json:"updated,omitempty"`
	UserId   *int32  `json:"user_id,omitempty"`
}

// GetApiV1RolesParams defines parameters for GetApiV1Roles.
type GetApiV1RolesParams struct {
	Name     *string `form:"name,omitempty" json:"name,omitempty"`
	Status   *string `form:"status,omitempty" json:"status,omitempty"`
	Current  int     `form:"current" json:"current"`
	PageSize int     `form:"pageSize" json:"pageSize"`
}

// GetApiV1UsersParams defines parameters for GetApiV1Users.
type GetApiV1UsersParams struct {
	Username *string `form:"username,omitempty" json:"username,omitempty"`
	Name     *string `form:"name,omitempty" json:"name,omitempty"`
	Status   *string `form:"status,omitempty" json:"status,omitempty"`
	Current  int     `form:"current" json:"current"`
	PageSize int     `form:"pageSize" json:"pageSize"`
}

// PostApiV1RolesJSONRequestBody defines body for PostApiV1Roles for application/json ContentType.
type PostApiV1RolesJSONRequestBody = Role

// PutApiV1RolesIdJSONRequestBody defines body for PutApiV1RolesId for application/json ContentType.
type PutApiV1RolesIdJSONRequestBody = Role

// PostApiV1UsersJSONRequestBody defines body for PostApiV1Users for application/json ContentType.
type PostApiV1UsersJSONRequestBody = User

// PutApiV1UsersIdJSONRequestBody defines body for PutApiV1UsersId for application/json ContentType.
type PutApiV1UsersIdJSONRequestBody = User